Moved stickers binding with Sample Types upgrade step to 1.2.2
